+++ gridsome/__tests__/schema.spec.js
@@ -1,27 +1,14 @@
 const path = require('path')
 const App = require('../lib/app/App')
+const { graphql } = require('../graphql')
 const PluginAPI = require('../lib/app/PluginAPI')
 const createSchema = require('../lib/graphql/createSchema')
-const { inferTypes } = require('../lib/graphql/schema/infer-types')
-const { GraphQLDate } = require('../lib/graphql/schema/types/date')
-const { fileType } = require('../lib/graphql/schema/types/file')
-const { imageType } = require('../lib/graphql/schema/types/image')
 const JSONTransformer = require('./__fixtures__/JSONTransformer')
 
 const context = __dirname
 const targetDir = path.join(context, 'assets', 'static')
 const assetsDir = path.join(targetDir, 'assets')
 const pathPrefix = '/'
-
-const {
-  graphql,
-  GraphQLInt,
-  GraphQLList,
-  GraphQLFloat,
-  GraphQLString,
-  GraphQLBoolean,
-  GraphQLObjectType
-} = require('../graphql')
 
 let app, api
 
@@ -36,7 +23,7 @@ beforeEach(() => {
       maxImageWidth: 1000
     }
   }).init()
-  
+
   api = new PluginAPI(app, {
     entry: { options: {}, clientOptions: undefined },
     transformers: {
@@ -70,7 +57,18 @@ test('create node type with custom fields', async () => {
     }
   })
 
-  const query = '{ testPost (id: "1") { id _id foo list obj { foo }}}'
+  contentType.addNode({
+    id: '2',
+    fields: {
+      foo: 'bar',
+      list: ['item'],
+      obj: {
+        foo: 'bar'
+      }
+    }
+  })
+
+  const query = '{ testPost (id: "1") { id foo list obj { foo }}}'
   const { data } = await createSchemaAndExecute(query)
 
   expect(data.testPost.id).toEqual('1')
@@ -160,6 +158,32 @@ test('create connection', async () => {
   expect(data.allTestPost.edges[1].node.title).toEqual('test 1')
 })
 
+test('sort nodes collection', async () => {
+  const contentType = api.store.addContentType({
+    typeName: 'TestPost'
+  })
+
+  contentType.addNode({ title: 'c' })
+  contentType.addNode({ title: 'b' })
+  contentType.addNode({ title: 'a' })
+
+  const query = `{
+    allTestPost (sortBy: "title", order: ASC) {
+      edges {
+        node { title }
+      }
+    }
+  }`
+
+  const { errors, data } = await createSchemaAndExecute(query)
+
+  expect(errors).toBeUndefined()
+  expect(data.allTestPost.edges.length).toEqual(3)
+  expect(data.allTestPost.edges[0].node.title).toEqual('a')
+  expect(data.allTestPost.edges[1].node.title).toEqual('b')
+  expect(data.allTestPost.edges[2].node.title).toEqual('c')
+})
+
 test('get nodes by path regex', async () => {
   const contentType = api.store.addContentType({
     typeName: 'TestPost'
@@ -190,14 +214,39 @@ test('create node reference', async () => {
     }
   })
 
-  postContentType.addNode({ id: '1', fields: { author: '2' }})
-  authorContentType.addNode({ id: '2', title: 'Test Author' })
+  authorContentType.addNode({
+    id: '2',
+    title: 'Test Author'
+  })
 
-  const query = '{ testPost (id: "1") { author { id title }}}'
+  postContentType.addNode({
+    id: '1',
+    fields: {
+      author: '2',
+      customRefs: {
+        author: {
+          typeName: 'TestAuthor',
+          value: '2'
+        }
+      }
+    }
+  })
+
+  const query = `{
+    testPost (id: "1") {
+      author { id title }
+      customRefs {
+        author { id title }
+      }
+    }
+  }`
+
   const { data } = await createSchemaAndExecute(query)
 
   expect(data.testPost.author.id).toEqual('2')
   expect(data.testPost.author.title).toEqual('Test Author')
+  expect(data.testPost.customRefs.author.id).toEqual('2')
+  expect(data.testPost.customRefs.author.title).toEqual('Test Author')
 })
 
 // TODO: remove this test before 1.0
@@ -241,14 +290,39 @@ test('create node list reference', async () => {
     }
   })
 
-  postContentType.addNode({ id: '1', fields: { author: ['2'] }})
-  authorContentType.addNode({ id: '2', title: 'Test Author' })
+  authorContentType.addNode({
+    id: '2',
+    title: 'Test Author'
+  })
 
-  const query = '{ testPost (id: "1") { refs { author { id title }}}}'
+  postContentType.addNode({
+    id: '1',
+    fields: {
+      author: ['2'],
+      customRefs: {
+        author: {
+          typeName: 'TestAuthor',
+          value: ['2']
+        }
+      }
+    }
+  })
+
+  const query = `{
+    testPost (id: "1") {
+      author { id title }
+      customRefs {
+        author { id title }
+      }
+    }
+  }`
+
   const { data } = await createSchemaAndExecute(query)
 
-  expect(data.testPost.refs.author[0].id).toEqual('2')
-  expect(data.testPost.refs.author[0].title).toEqual('Test Author')
+  expect(data.testPost.author[0].id).toEqual('2')
+  expect(data.testPost.author[0].title).toEqual('Test Author')
+  expect(data.testPost.customRefs.author[0].id).toEqual('2')
+  expect(data.testPost.customRefs.author[0].title).toEqual('Test Author')
 })
 
 test('create node reference to same type', async () => {
@@ -269,105 +343,6 @@ test('create node reference to same type', async () => {
 
   expect(data.testPost.refs.related.id).toEqual('2')
   expect(data.testPost.refs.related.title).toEqual('Test')
-})
-
-test('infer types from node fields', () => {
-  const types = inferTypes([
-    {
-      fields: {
-        string: 'bar',
-        number: 10,
-        float: 1.2,
-        truthyBoolean: true,
-        stringList: ['item'],
-        numberList: [10],
-        floatList: [1.2],
-
-        obj: {
-          foo: 'bar'
-        }
-      }
-    },
-    {
-      fields: {
-        string: null,
-        falsyBoolean: false,
-        booleanList: [false],
-        emptyList: [],
-        emptyObj: {},
-        emptyString: ''
-      }
-    }
-  ], 'TestPost')
-
-  expect(types.string.type).toEqual(GraphQLString)
-  expect(types.number.type).toEqual(GraphQLInt)
-  expect(types.float.type).toEqual(GraphQLFloat)
-  expect(types.falsyBoolean.type).toEqual(GraphQLBoolean)
-  expect(types.truthyBoolean.type).toEqual(GraphQLBoolean)
-  expect(types.stringList.type).toBeInstanceOf(GraphQLList)
-  expect(types.stringList.type.ofType).toEqual(GraphQLString)
-  expect(types.numberList.type.ofType).toEqual(GraphQLInt)
-  expect(types.floatList.type.ofType).toEqual(GraphQLFloat)
-  expect(types.booleanList.type.ofType).toEqual(GraphQLBoolean)
-  expect(types.obj.type).toBeInstanceOf(GraphQLObjectType)
-  expect(types.obj.type.name).toEqual('TestPostObj')
-  expect(types.obj.type.getFields().foo.type).toEqual(GraphQLString)
-  expect(types.emptyList).toBeUndefined()
-  expect(types.emptyObj).toBeUndefined()
-  expect(types.emptyString).toBeUndefined()
-})
-
-test('infer date fields', () => {
-  const types = inferTypes([
-    {
-      fields: {
-        date1: '2018',
-        date2: '2018-11',
-        date3: '2018-11-01',
-        date4: '2018-11-01T19:20+01:00',
-        date5: '2018-11-01T19:20:30+01:00'
-      }
-    },
-  ], 'TestPost')
-
-  expect(types.date1.type).toEqual(GraphQLDate)
-  expect(types.date2.type).toEqual(GraphQLDate)
-  expect(types.date3.type).toEqual(GraphQLDate)
-  expect(types.date4.type).toEqual(GraphQLDate)
-  expect(types.date5.type).toEqual(GraphQLDate)
-})
-
-test('infer image fields', () => {
-  const types = inferTypes([
-    {
-      fields: {
-        image1: 'image.png',
-        image2: '/image.png',
-        image3: './image.png',
-        image4: 'https://www.example.com/images/image.png'
-      }
-    },
-  ], 'TestPost')
-
-  expect(types.image1.type).toEqual(imageType.type)
-  expect(types.image2.type).toEqual(imageType.type)
-  expect(types.image3.type).toEqual(imageType.type)
-  expect(types.image4.type).toEqual(imageType.type)
-})
-
-test('infer file fields', () => {
-  const types = inferTypes([
-    {
-      fields: {
-        file1: './document.pdf',
-        file2: 'https://www.example.com/files/document.pdf'
-      }
-    },
-  ], 'TestPost')
-
-  expect(types.file1.type).toEqual(fileType.type)
-  expect(types.file2.type).toEqual(fileType.type)
 })
 
 test('should get values from object fields', async () => {
@@ -544,7 +519,10 @@ test('process file types in schema', async () => {
       content: JSON.stringify({
         file: '/assets/document.pdf',
         file2: 'https://www.example.com/assets/document.pdf',
-        file3: './dummy.pdf'
+        file3: './dummy.pdf',
+        url: 'https://www.gridsome.org',
+        url2: 'https://www.gridsome.com',
+        text: 'pdf'
       })
     }
   })
@@ -554,6 +532,9 @@ test('process file types in schema', async () => {
       file
       file2
       file3
+      url
+      url2
+      text
     }
   }`)
 
@@ -566,6 +547,9 @@ test('process file types in schema', async () => {
   expect(data.testPost.file3.type).toEqual('file')
   expect(data.testPost.file3.mimeType).toEqual('application/pdf')
   expect(data.testPost.file3.src).toEqual('/assets/files/dummy.pdf')
+  expect(data.testPost.url).toEqual('https://www.gridsome.org')
+  expect(data.testPost.url2).toEqual('https://www.gridsome.com')
+  expect(data.testPost.text).toEqual('pdf')
 })
 
 async function createSchemaAndExecute (query) {